Add button to logout into dropdown
