+++ frontend/src/pages/dashboard/Dashboard.tsx
@@ -1,31 +1,20 @@
 import React from 'react';
-import signout from '../../hooks/signout';
 import StudentDashboard from '../../components/courses/Students';
 import { useAuthContext } from '../../hooks/useAuthContext';
 import { useNavigate } from 'react-router-dom';
 import DashboardWrapper from './DashboardWrapper';
 
 const Dashboard: React.FC = () => {
-  const { user, setUser } = useAuthContext();
+  const { user } = useAuthContext();
   const navigate = useNavigate();
 
   if (!user?.emailVerified) {
     navigate('/login', { replace: true });
   }
 
-  const handleLogout = async (): Promise<void> => {
-    setUser(null);
-    await signout();
-    navigate("/login", { replace: true })
-  };
-
   return (
     <DashboardWrapper>
-    <div>
-      <h1>Hey {user?.displayName}, Welcome to the e-learning-platform.</h1>
-      <button onClick={handleLogout}>Logout</button>
       <StudentDashboard />
-    </div>
     </DashboardWrapper>
   );
 };
